make MC/SA fortran stuff all REAL*8 add glFreeType.py - doesn't work correctly yet implement glFreeType font in OpenGl labels add new menu items for RBs MCSA anneal in a test mode no random start for fixed T0
git-svn-id: https://subversion.xray.aps.anl.gov/pyGSAS/trunk@978 2b3fe19a-dc8f-cc80-9b5f-a4fb0fcdb5be

+++ fsource/mcsasubs/prodvmv.for
@@ -4,7 +4,7 @@
 !
 ! ROUTINE TO COMPUTE THE DOT PRODUCT OF TWO 3 VECTORS
 
-      REAL          X(3),Y(3),A(3,3),PRODVMV 
+      REAL*8        X(3),Y(3),A(3,3),PRODVMV 
 
       T1 = 0.0
       DO J=1,3
